refactor(frontend/orders): modularize UI and logic, standardize CSS
- Rename page-specific CSS files to page.module.css
- Extract complex UI sections to _components/ (CheckoutContent, CancelDialog)
- Extract state/fetch logic into dedicated custom hooks next to the components
- Add Order_FE_Refactoring_Rules.md documentation for conventions
