+++ docs/Architecture/Order_FE_Refactoring_Rules.md
@@ -1,0 +1,29 @@
+# 프론트엔드 리팩토링 규칙 (Order 모듈 기반)
+
+최근 진행된 `orders` 모듈 리팩토링에서 준수한 핵심 설계 지침 및 규칙을 정리한 문서입니다. 향후 프로젝트 내 다른 페이지나 모듈 코드를 작성/리팩토링할 때도 동일한 기준을 적용합니다.
+
+## 1. CSS 파일명 및 구조 통일
+- **규칙**: 컴포넌트나 루트 페이지에 1:1로 매칭되는 CSS 모듈 파일명을 일관성 있게 생성합니다.
+- **적용 패턴**: 모든 페이지 렌더링에 사용되는 최상위 CSS는 `page.module.css`로 통일했습니다. (예: `checkout.module.css` 제거 및 해당 폴더 하위의 `page.module.css`로 대체)
+- **이유**: Next.js App Router의 라우팅 구조(page.tsx)와 직관적으로 맞추어, 누구나 파일 구조를 파악하기 쉽게 만듭니다.
+
+## 2. 컴포넌트 간 CSS 독립성(캡슐화) 완전 보장
+- **규칙**: 다른 페이지나 부모/형제 폴더에 있는 CSS 파일을 종속적으로 `import` 하여 재사용하지 않습니다.
+- **적용 패턴**: `success/page.tsx`, `fail/page.tsx` 등에서 상위의 `checkout.module.css`를 참조하여 사용하던 것을 제거했습니다. 공통으로 필요한 레이아웃/스타일이라도 자신의 폴더 내 로컬 `page.module.css`에 직접 선언하여 의존성을 끊었습니다.
+- **이유**: 특정 컴포넌트의 스타일 코드가 변경되었을 때 엉뚱한 타 화면 레이아웃이 무너지는 잠재적 사이드 이펙트를 완벽히 방지하기 위함입니다.
+
+## 3. UI 렌더링과 비즈니스/상태 로직의 철저한 분리 (Custom Hook)
+- **규칙**: 무거운 상태(`useState`), API 응답 사이드 이펙트 관리(`useEffect`), Fetch 통신 로직 등은 모두 커스텀 훅(Custom Hook)으로 추출합니다.
+- **적용 패턴**: `useCheckout.ts`, `usePaymentConfirm.ts`, `useRefund.ts`처럼 각각의 뷰 컴포넌트에 1:1로 대응되는 비즈니스 로직 훅을 생성했습니다.
+- **위치 정책**: 생성된 커스텀 훅은 불필요한 별도 계층(`_hooks/` 등 공용 폴더)으로 빼지 않고, **실제 이 훅을 호출하는 컴포넌트와 동일한 폴더 구조에 인접시켜** 코드의 응집도를 높였습니다.
+- **이유**: `page.tsx` 또는 `View Component`는 '화면을 어떻게 그릴 것인가'라는 표현(Presentation) 역할에만 집중시키는 단일 책임 원칙(SRP)을 준수합니다.
+
+## 4. 관심사 모듈 세분화 (Sub-components 추출)
+- **규칙**: `page.tsx` 하나에 모달, 다이얼로그, 내부 탭, 폼 영역 등이 모두 포함되어 수백 줄 이상의 방대한 코드가 되는 것을 지양합니다.
+- **적용 패턴**: `CancelDialog.tsx`, `CheckoutContent.tsx` 등 독립적으로 묶일 수 있는 영역은 `_components/` 폴더 내로 추출했으며, 연결된 전용 스타일 구문들도 컴포넌트에 맞게 분리(`CancelDialog.module.css` 형태)하여 이동시켰습니다.
+- **이유**: `page.tsx`의 볼륨을 최소화시켜 진입점(Entry Point)이자 래퍼로서의 역할만 하독록 강제시켜, 코드 가독성과 유지보수성을 극대화합니다.
+
+## 5. 복합 컴포넌트의 폴더화 (Component Directory Pattern)
+- **규칙**: 단일 컴포넌트 기능 구현을 위해 관련된 `.tsx` 파일이 2개 이상 생성되는 경우, 내부 하위 폴더를 생성하여 그룹화합니다.
+- **적용 패턴**: 단일 `_components/ComponentA.tsx` 파일 내에서 관리하기 벅차 분할이 필요한 경우, `_components/ComponentA/` 폴더를 생성하고 그 안에 관련된 `.tsx` 파일들과 전용 `.module.css`를 함께 응집시킵니다.
+- **이유**: 폴더 내부 구조가 지저분해지는 것을 막고 하나의 컴포넌트 단위를 통째로 관리 가능한 패키지/모듈처럼 작동하게 합니다.
