+++ frontend/src/app/orders/checkout/fail/page.module.css
@@ -43,3 +43,29 @@
   width: 100%;
   justify-content: center;
 }
+.checkoutContainer {
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  min-height: 100vh;
+  padding: var(--space-48) var(--space-24);
+  width: 100%;
+}
+
+.checkoutCard {
+  background-color: var(--color-bg);
+  width: 100%;
+  max-width: 600px;
+  margin: 0 auto;
+  border-radius: var(--radius-xl);
+  border: 1px solid var(--color-gray-disabled-fill);
+  padding: var(--space-32);
+  box-shadow: var(--shadow-lg);
+}
+
+@media (max-width: 768px) {
+  .checkoutCard {
+    padding: var(--space-24) var(--space-16);
+    border-radius: var(--radius-lg);
+  }
+}
