+++ frontend/src/app/orders/checkout/success/page.module.css
@@ -1,4 +1,29 @@
-/* 피그마 디자인 기반 성공 페이지 레이아웃 */
+.checkoutContainer {
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  min-height: 100vh;
+  padding: var(--space-48) var(--space-24);
+  width: 100%;
+}
+
+.checkoutCard {
+  background-color: var(--color-bg);
+  width: 100%;
+  max-width: 600px;
+  margin: 0 auto;
+  border-radius: var(--radius-xl);
+  border: 1px solid var(--color-gray-disabled-fill);
+  padding: var(--space-32);
+  box-shadow: var(--shadow-lg);
+}
+
+@media (max-width: 768px) {
+  .checkoutCard {
+    padding: var(--space-24) var(--space-16);
+    border-radius: var(--radius-lg);
+  }
+}/* 피그마 디자인 기반 성공 페이지 레이아웃 */
 .pageContainer {
   display: flex;
   flex-direction: column;
