+++ frontend/src/app/orders/checkout/success/page.tsx
@@ -1,121 +1,53 @@
 "use client";
 
-import React, { Suspense, useEffect, useState } from 'react';
+import React, { Suspense } from 'react';
 import { useSearchParams } from 'next/navigation';
-import styles from '../checkout.module.css';
-import successStyles from './success.module.css';
+import styles from './page.module.css';
 import { Button } from '@/components/ui';
+import { usePaymentConfirm } from './usePaymentConfirm';
 
 
 function CheckoutSuccessContent() {
   const searchParams = useSearchParams();
-  const [status, setStatus] = useState<'loading' | 'success' | 'error'>('loading');
-  const [message, setMessage] = useState('결제를 승인하는 중입니다...');
-  const [paymentInfo, setPaymentInfo] = useState<any>(null);
-
-  useEffect(() => {
-    const confirmPayment = async () => {
-      const paymentKey = searchParams.get('paymentKey');
-      const tossOrderId = searchParams.get('orderId'); // 토스가 보낸 결제 식별값 (venueon_batch_...)
-      const amount = searchParams.get('amount');
-      const backendOrderId = searchParams.get('backendOrderId'); // 우리가 보낸 내부 DB ID (숫자)
-
-      if (!paymentKey || !tossOrderId || !amount) {
-        setStatus('error');
-        setMessage('결제 정보가 올바르지 않습니다.');
-        return;
-      }
-
-      // 내부 orderId 식별 (쿼리 파라미터 backendOrderId 최우선)
-      const match = tossOrderId.match(/venueon_order_(\d+)_/);
-      let extractedOrderId = backendOrderId || (match ? match[1] : null);
-
-      if (!extractedOrderId) {
-        setStatus('error');
-        setMessage('주문 ID를 확인할 수 없습니다.');
-        return;
-      }
-
-      try {
-        const res = await fetch(`/api/orders/${extractedOrderId}/confirm`, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            paymentKey,
-            orderId: tossOrderId,
-            amount: Number(amount),
-          }),
-        });
-
-        if (!res.ok) {
-          const errorBody = await res.json().catch(() => null);
-          throw new Error(errorBody?.message || `결제 승인 실패 (${res.status})`);
-        }
-
-        const json = await res.json();
-        setPaymentInfo(json.data);
-        setStatus('success');
-        setMessage('결제가 완료되었습니다!');
-      } catch (err: any) {
-        console.error('결제 승인 에러:', err);
-        setStatus('error');
-
-        let friendlyMessage = '결제 승인 중 오류가 발생했습니다.';
-        const rawError = err.message || '';
-
-        // 백엔드에서 뱉는 형태: "토스 결제 승인 실패: 400 Bad Request: "{...}"" 에서 메시지 추출
-        try {
-          const messageMatch = rawError.match(/"message":"([^"]+)"/);
-          if (messageMatch && messageMatch[1]) {
-            friendlyMessage = messageMatch[1];
-          } else if (rawError.includes('ALREADY_PROCESSED_PAYMENT')) {
-            friendlyMessage = '이미 처리된 결제입니다.';
-          } else {
-            // 다른 텍스트만 있다면 그대로 노출 방지(안전을 위해 일반화)
-          }
-        } catch (e) {
-          // 파싱 실패시 기본 메시지
-        }
-
-        setMessage(friendlyMessage);
-      }
-    };
-
-    confirmPayment();
-  }, [searchParams]);
+  const { status, message, paymentInfo } = usePaymentConfirm(
+    searchParams.get('paymentKey'),
+    searchParams.get('orderId'),
+    searchParams.get('amount'),
+    searchParams.get('backendOrderId')
+  );
 
   const formatPrice = (price: number) =>
     new Intl.NumberFormat('ko-KR').format(price) + '원';
 
   if (status === 'success') {
     return (
-      <div className={successStyles.pageContainer}>
-        <div className={successStyles.headerSection}>
-          <h1 className={successStyles.title}>수강 신청 완료!</h1>
-          <p className={successStyles.subtitle}>
+      <div className={styles.pageContainer}>
+        <div className={styles.headerSection}>
+          <h1 className={styles.title}>수강 신청 완료!</h1>
+          <p className={styles.subtitle}>
             교육 프로그램 등록이 완료되었습니다.<br />
             상세 내역은 내 세션 목록에서 확인하실 수 있습니다.
           </p>
         </div>
 
         {paymentInfo && (
-          <div className={successStyles.cardSection}>
-            <div className={successStyles.cardRow}>
-              <p className={successStyles.cardLabel}>주문 번호</p>
-              <p className={successStyles.cardValue}>#{paymentInfo.orderId}</p>
+          <div className={styles.cardSection}>
+            <div className={styles.cardRow}>
+              <p className={styles.cardLabel}>주문 번호</p>
+              <p className={styles.cardValue}>#{paymentInfo.orderId}</p>
             </div>
-            <div className={successStyles.cardRow}>
-              <p className={successStyles.cardLabel}>세션명</p>
-              <p className={successStyles.cardValue}>{paymentInfo.orderName || 'VenueOn 세션'}</p>
+            <div className={styles.cardRow}>
+              <p className={styles.cardLabel}>세션명</p>
+              <p className={styles.cardValue}>{paymentInfo.orderName || 'VenueOn 세션'}</p>
             </div>
-            <div className={successStyles.cardRow}>
-              <p className={successStyles.cardLabel}>결제 금액</p>
-              <p className={successStyles.cardValue}>{formatPrice(paymentInfo.amount)}</p>
+            <div className={styles.cardRow}>
+              <p className={styles.cardLabel}>결제 금액</p>
+              <p className={styles.cardValue}>{formatPrice(paymentInfo.amount)}</p>
             </div>
           </div>
         )}
 
-        <div className={successStyles.buttonGroup}>
+        <div className={styles.buttonGroup}>
           <Button
             variant="outlined"
             size="large"
@@ -135,11 +67,11 @@ function CheckoutSuccessContent() {
         </div>
 
         {/* 우측 하단 토스트 모달 (필요시 유지) */}
-        <div className={successStyles.toastContainer}>
-          <div className={successStyles.toastIcon}>🎉</div>
-          <div className={successStyles.toastContent}>
-            <h4 className={successStyles.toastTitle}>수강신청 완료!</h4>
-            <p className={successStyles.toastMessage}>성공적으로 등록되었습니다.</p>
+        <div className={styles.toastContainer}>
+          <div className={styles.toastIcon}>🎉</div>
+          <div className={styles.toastContent}>
+            <h4 className={styles.toastTitle}>수강신청 완료!</h4>
+            <p className={styles.toastMessage}>성공적으로 등록되었습니다.</p>
           </div>
         </div>
       </div>
