+++ frontend/src/app/orders/refund/page.module.css
@@ -140,137 +140,6 @@
   margin-bottom: 12px;
 }
 
-/* --- 취소 사유 다이얼로그 --- */
-.overlay {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.45);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000;
-  animation: fadeIn 0.2s ease;
-}
-
-.dialog {
-  background: white;
-  border-radius: 16px;
-  padding: 32px;
-  width: 100%;
-  max-width: 440px;
-  box-shadow: 0 25px 50px rgba(0, 0, 0, 0.15);
-  animation: slideUp 0.3s ease;
-}
-
-.dialogTitle {
-  font-size: 20px;
-  font-weight: 700;
-  color: var(--color-gray-900, #111827);
-  margin: 0 0 8px 0;
-}
-
-.dialogSubtitle {
-  font-size: 14px;
-  color: var(--color-gray-500, #6b7280);
-  margin: 0 0 20px 0;
-}
-
-.reasonOptions {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  margin-bottom: 16px;
-}
-
-.reasonOption {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 12px 16px;
-  border-radius: 10px;
-  border: 1px solid var(--color-gray-200, #e5e7eb);
-  cursor: pointer;
-  transition: all 0.15s ease;
-  font-size: 14px;
-  color: var(--color-gray-900, #111827);
-  background: white;
-}
-
-.reasonOption:hover {
-  border-color: var(--color-gray-400, #9ca3af);
-}
-
-.reasonOptionSelected {
-  border-color: var(--color-gray-900, #111827);
-  background: var(--color-gray-50, #f9fafb);
-}
-
-.reasonRadio {
-  width: 18px;
-  height: 18px;
-  accent-color: var(--color-gray-900, #111827);
-}
-
-.customReasonInput {
-  width: 100%;
-  padding: 12px 16px;
-  border-radius: 10px;
-  border: 1px solid var(--color-gray-200, #e5e7eb);
-  font-size: 14px;
-  margin-bottom: 20px;
-  outline: none;
-  transition: border-color 0.15s ease;
-  box-sizing: border-box;
-}
-
-.customReasonInput:focus {
-  border-color: var(--color-gray-900, #111827);
-}
-
-.dialogActions {
-  display: flex;
-  gap: 12px;
-}
-
-.dialogCancelBtn {
-  flex: 1;
-  padding: 12px;
-  border-radius: 10px;
-  border: 1px solid var(--color-gray-200, #e5e7eb);
-  background: white;
-  color: var(--color-gray-700, #374151);
-  font-size: 15px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background 0.15s ease;
-}
-
-.dialogCancelBtn:hover {
-  background: var(--color-gray-50, #f9fafb);
-}
-
-.dialogConfirmBtn {
-  flex: 1;
-  padding: 12px;
-  border-radius: 10px;
-  border: none;
-  background: #DC2626;
-  color: white;
-  font-size: 15px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background 0.15s ease;
-}
-
-.dialogConfirmBtn:hover {
-  background: #B91C1C;
-}
-
-.dialogConfirmBtn:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
 /* 토스트 */
 .toast {
   position: fixed;
