+++ frontend/src/app/orders/refund/page.tsx
@@ -1,96 +1,20 @@
 "use client";
 
-import React, { useEffect, useState } from 'react';
-import styles from './refund.module.css';
-
-// 취소 사유 선택지
-const CANCEL_REASONS = [
-  '일정이 변경되어 참석이 어렵습니다.',
-  '다른 세션를 수강하고 싶습니다.',
-  '단순 변심입니다.',
-  '세션 내용이 기대와 달랐습니다.',
-  '직접 입력',
-];
-
-interface OrderItem {
-  orderId: number;
-  eventId: number;
-  eventTitle: string;
-  status: string;
-  quantity: number;
-  amount: number;
-  paymentMethod: string;
-  orderedAt: string;
-  paidAt: string | null;
-}
+import React from 'react';
+import styles from './page.module.css';
+import CancelDialog from './_components/CancelDialog';
+import { useRefund } from './useRefund';
 
 export default function RefundPage() {
-  const [orders, setOrders] = useState<OrderItem[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [cancelTarget, setCancelTarget] = useState<OrderItem | null>(null);
-  const [selectedReason, setSelectedReason] = useState('');
-  const [customReason, setCustomReason] = useState('');
-  const [cancelling, setCancelling] = useState(false);
-  const [toast, setToast] = useState<{ title: string; message: string } | null>(null);
-
-  // 주문 목록 불러오기
-  const fetchOrders = async () => {
-    try {
-      const res = await fetch('/api/orders/me?page=0&size=50');
-      if (!res.ok) throw new Error('주문 목록을 불러올 수 없습니다.');
-      const json = await res.json();
-      // Page 형태의 응답 파싱
-      const content = json.data?.content || json.data || [];
-      setOrders(Array.isArray(content) ? content : []);
-    } catch (err) {
-      console.error('주문 목록 조회 실패:', err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchOrders();
-  }, []);
-
-  // 취소 요청 실행
-  const handleCancelConfirm = async () => {
-    if (!cancelTarget) return;
-
-    const reason = selectedReason === '직접 입력' ? customReason : selectedReason;
-    if (!reason.trim()) return;
-
-    setCancelling(true);
-    try {
-      const res = await fetch(`/api/orders/${cancelTarget.orderId}/refund`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ reason }),
-      });
-
-      if (!res.ok) {
-        const errorBody = await res.json().catch(() => null);
-        throw new Error(errorBody?.message || '환불 요청에 실패했습니다.');
-      }
-
-      // 성공: 로컬 목록에서 해당 주문의 상태를 즉시 REFUNDED로 갱신
-      setOrders((prev) =>
-        prev.map((o) =>
-          o.orderId === cancelTarget.orderId ? { ...o, status: 'REFUNDED' } : o
-        )
-      );
-      setCancelTarget(null);
-      setSelectedReason('');
-      setCustomReason('');
-      setToast({ title: '환불 완료', message: `${cancelTarget.eventTitle} 수강이 취소되었습니다.` });
-      setTimeout(() => setToast(null), 4000);
-
-    } catch (err: any) {
-      alert(err.message || '환불 처리 중 오류가 발생했습니다.');
-    } finally {
-      setCancelling(false);
-    }
-  };
+  const {
+    orders,
+    loading,
+    cancelling,
+    cancelTarget,
+    setCancelTarget,
+    toast,
+    handleCancelConfirm
+  } = useRefund();
 
   // 가격 포맷
   const formatPrice = (price: number) =>
@@ -160,8 +84,6 @@ export default function RefundPage() {
                     className={styles.cancelBtn}
                     onClick={() => {
                       setCancelTarget(order);
-                      setSelectedReason('');
-                      setCustomReason('');
                     }}
                   >
                     수강 취소
@@ -175,64 +97,12 @@ export default function RefundPage() {
 
       {/* 취소 사유 다이얼로그 */}
       {cancelTarget && (
-        <div className={styles.overlay} onClick={() => !cancelling && setCancelTarget(null)}>
-          <div className={styles.dialog} onClick={(e) => e.stopPropagation()}>
-            <h2 className={styles.dialogTitle}>수강 취소</h2>
-            <p className={styles.dialogSubtitle}>
-              <strong>{cancelTarget.eventTitle}</strong>의 수강을 취소하시겠어요?<br />
-              취소 사유를 선택해 주세요.
-            </p>
-
-            <div className={styles.reasonOptions}>
-              {CANCEL_REASONS.map((reason) => (
-                <label
-                  key={reason}
-                  className={`${styles.reasonOption} ${selectedReason === reason ? styles.reasonOptionSelected : ''}`}
-                >
-                  <input
-                    type="radio"
-                    name="cancelReason"
-                    className={styles.reasonRadio}
-                    checked={selectedReason === reason}
-                    onChange={() => setSelectedReason(reason)}
-                  />
-                  {reason}
-                </label>
-              ))}
-            </div>
-
-            {selectedReason === '직접 입력' && (
-              <input
-                type="text"
-                className={styles.customReasonInput}
-                placeholder="취소 사유를 직접 입력해 주세요"
-                value={customReason}
-                onChange={(e) => setCustomReason(e.target.value)}
-              />
-            )}
-
-            <div className={styles.dialogActions}>
-              <button
-                className={styles.dialogCancelBtn}
-                onClick={() => setCancelTarget(null)}
-                disabled={cancelling}
-              >
-                돌아가기
-              </button>
-              <button
-                className={styles.dialogConfirmBtn}
-                onClick={handleCancelConfirm}
-                disabled={
-                  cancelling ||
-                  !selectedReason ||
-                  (selectedReason === '직접 입력' && !customReason.trim())
-                }
-              >
-                {cancelling ? '처리 중...' : '환불 요청'}
-              </button>
-            </div>
-          </div>
-        </div>
+        <CancelDialog
+          cancelTarget={cancelTarget}
+          cancelling={cancelling}
+          onConfirm={handleCancelConfirm}
+          onClose={() => setCancelTarget(null)}
+        />
       )}
 
       {/* 토스트 알림 */}
